Fixed the parser for the percentils in the c3 d3.js

+++ emif/population_characteristics/static/js/emif.charts.c3d3.js
@@ -78,7 +78,7 @@ function GraphicChartC3D3(divArg, dataArg)
 
           
 
-          datasetYs[k].push(parseInt(row[a.trim()]));  
+          datasetYs[k].push(parseFloat(row[a.trim()]));  
           k = k +1 ;
         });
